// cocos_client2D/assets/scripts/Generated/Model/demo/protocol/messageDescriptors.ts
// Generated by tools/codegen_proto.mjs. Do not edit by hand.
import { defineMessage } from "../../../../Core/Protocol/Message";
import {
  C2M_Move,
  C2M_MoveCodec,
  G2C_EntityEnter,
  G2C_EntityEnterCodec,
  G2C_EntityLeave,
  G2C_EntityLeaveCodec,
  G2C_EntityMove,
  G2C_EntityMoveCodec,
  G2C_MapReady,
  G2C_MapReadyCodec,
} from "./messages";
import { MsgCode } from "./msgcodes";

export const ClientMessages = {
  MapReady: defineMessage<G2C_MapReady>({
    name: "Client.MapReady",
    msgcode: MsgCode.G2C_MapReady,
    codec: G2C_MapReadyCodec,
  }),
  EntityMove: defineMessage<G2C_EntityMove>({
    name: "Client.EntityMove",
    msgcode: MsgCode.G2C_EntityMove,
    codec: G2C_EntityMoveCodec,
  }),
  EntityEnter: defineMessage<G2C_EntityEnter>({
    name: "Client.EntityEnter",
    msgcode: MsgCode.G2C_EntityEnter,
    codec: G2C_EntityEnterCodec,
  }),
  EntityLeave: defineMessage<G2C_EntityLeave>({
    name: "Client.EntityLeave",
    msgcode: MsgCode.G2C_EntityLeave,
    codec: G2C_EntityLeaveCodec,
  }),
};

export const MapMessages = {
  Move: defineMessage<C2M_Move>({
    name: "Map.Move",
    msgcode: MsgCode.C2M_Move,
    codec: C2M_MoveCodec,
    routing: "actor-location",
  }),
};

export const AllMessageDescriptors = [
  ClientMessages.MapReady,
  MapMessages.Move,
  ClientMessages.EntityMove,
  ClientMessages.EntityEnter,
  ClientMessages.EntityLeave,
] as const;