// cocos_client2D/assets/scripts/Generated/Model/demo/protocol/messages.ts
export interface MapEntitySnapshot {
  unitId: number;
  x: number;
  y: number;
  heading: number;
  alive: boolean;
  state: Uint8Array;
  account: string;
  cellX: number;
  cellY: number;
}